Delete secrets when downloading
I've decided to delete a secret on the server whether it successfully
decrypts or not. I'm not sure what good it'd be to keep it on the
server if it's failed to decrypt?

Also, if we fail to either decode the metadata (for the UUID), or to
actually delete the secret on the server I've opted to log this.